Loan default prediction using machine learning (Decision Trees, Native Bayes and Random Forest Algorithms)
  Financial institutions are faced daily with high rates of non-performing loans because of the inability to accurately predict defaulters before the disbursement of loans, this has led to many failures and the eventual collapse of some financial industries, particularly the banking industries. The bank’s profitability depends on how well it’s able to minimize nonperforming loans called bad debt. Through effective credit risk management, banks can predict, analyze, and mitigate the possibility of non-performing loans, which would greatly improve their profitability.  Predicting loan repayment in the past has been a very exhausting task which has been greatly done through traditional methods such as calculated risk analysis through credit scoring and risk profiling methods, this has not effectively prevented and predicted risk because of the dynamic nature of defaulters. 
The objective of this project is to reduce the level of non-performing loans in the banking industries, by effectively predicting the probability of default using past customer records such as personal income, loan amount, interest rate, default rate, loan purpose, loan intent, loan status etc, to give a meaningful insight using different machine learning algorithm for proper decision making. This model would serve as a guide to determine customers' behavior toward loan repayment and further categorizes customers into defaulter and non-defaulter by finding the relationship between each independent variable. 
For business continuity, a proper understanding of business models through data analysis is imperative for business sustainability and competitive advantage. Financial institutions can significantly maximize profit and establish a good borrower to lenders relationship by mitigating credit risk using big data. The traditional method of taking calculated risk from customer profiling can be greatly eradicated with the emergence of big data technology using machine learning algorithms. The term "machine learning techniques" refers to a variety of software tools that may turn data sets into "models," which are able to represent the data set and generalize it in order to make predictions about new data. Machine learning models are generally classified into dependent variables and independent variables. In general, the independent variables are the many elements that influence the outcome of the prediction and are used as the model's input. The dependent variable refers to the target of the prediction and is used as the output of the model. The result showed the different machine learning outputs for the various algorithm such as Decision tree, Random Forest, and Naïve Bayes and suggested the best algorithm based on its accuracy, recall, and precision. However, there is no generally best model for the algorithm. The different models are expected to perform differently based on the type of input dataset.
A data pipeline is used to explain the methods, which shows the sequence of steps in arriving at a successful prediction result. Credit risk management is a classification problem where the financial institution is determining whether customers would default or not in the payment of a loan. It’s either the customer defaults which is denoted with 1 or is a non-defaulter called 0 in the machine learning sequence. Big data can be used to predict a long range of problems such as cancer detection, virus attack detection, and financial prediction. The dataset is first stored in the HDFS to enable the machine learning process, then a comma-separated variable file (CSV) is imported on the spark environment using jupyter notebook, where further processing is done. A series of data preprocessing was done using Apache spark as the processing engine.  The modeling was done using machine learning algorithms such as Decision Tree, Random Forest, and Naïve Bayes on a spark library such as Pandas for cleaning the data, NumPy used for mathematical calculations, and seaborn used for visualization. The proposed method would be used for early prediction of bank customers defaulting in repayment of loans based on machine learning algorithms and ensemble learning techniques. The model is composed of the following stages, Data Understanding, Data preparation, Data loading, Machine learning modeling, and evaluation.
The classification technique would involve using machine learning algorithms that would learn how to assign classes from a problem domain into a class model. The model used for my classification technique is a multi-class classification that involves algorithms such as decision trees, Naïve Bayes, and Random Forest. A supervised learning model was used to train the model where the label class was known. A model training process was used from sklearn.model_selection called the train_test_split model to train and test the dataset. The first step in the machine learning process is to arrange the data into an acceptable format using the one-hot encoder which converts the categorical variable into a numerical array of numbers, then the model was further trained using the train_test_split model. In Pyspark, the library used to import the machine learning algorithm is called the Scikit-learn where the data is separated into the “features” and “Targets”. Decision tree, Random forest and Naïve Bayes algorithm are deployed from python’s scikit-learn library using pyspark. The algorithm can run on HDFS successfully through its remote access using the internal IP of the Master node. The data set was split into two pieces namely, the Training set and the testing set, by randomly selecting the samples of data without replacement. A ratio of 70% (0.7) of the rows would be put into my training set, and the remaining 30% (0.3) would be put into my test set ( “X-train,” “X-test”, “Y-train”, “y-test”). The Train model involves building and training the model on the predictive model. It is denoted with “X_train” and “y_train”). Then the model was tested with (“X_test” and “y_test”) and the performance is evaluated using the machine learning algorithm such as the Accuracy, Precision, Recall, and F1-score. 
The dataset was gotten from Kaggle.com.  I imported my dataset from my HDFS, the result shows  390,984 observations with 12 rows and 32,582 columns. Some preprocessing steps were used to remove attributes of extreme values. The confusion matrix shows the classification of the model defaulters where denoted with non-defaulter called 0  and defaulter with 1. The TP( True positive) and TN (True Negative) were taken for the prediction where the decision tree had a defaulter of 6243 and a non-defaulter of 1373. the random forest had a defaulter of 6725 and a non-defaulter of 1286. The naïve bayes algorithm gave a defaulter of 67 and non-defaulter of 1802.  The Random Forest ensemble machine learning techniques gave a better result than the rest of the models. The Random Forest technique had an accuracy of 93%, an F1 score of 96%, a precision of 97%, and a recall of 99% as compared to other results from the machine learning algorithm. It shows that the ensemble machine-learning techniques had a better performance according to the result shown above.